# litedram/address_mapping Add bank_byte_alignment